Add files via upload
Task 6: Write a function that accepts a string. Convert the string to an array of words and remove the last word using pop().
Task 7: Create an array of days of the week. Using shift() and unshift(), move Sunday to the end of the array.
Task 8: Given an array of letters, write a function that adds a letter ‘Z’ at the beginning and end of the array.
Task 9: Initialize an array with five movie names. Ask the user for another movie name and add it to the end of the array.
Task 10: Remove the third item from the movie array